refactor(tunnel-lib): reorganize root-level modules into subdirs per architecture guidelines
- shared.rs → models/defs.rs (domain data types belong in models/)
- ctld_proto.rs → protocol/ctld.rs (wire protocol codec is protocol layer)
- inflight.rs + overload.rs → lb/ (load-balancing/backpressure primitives)
- accept.rs + open_bi.rs → transport/ (connection primitives belong in transport/)
- sniff.rs → protocol/sniff.rs (protocol detection belongs in protocol/)
- lib.rs re-exports all types at original pub paths (backward compat)
- Add ctld_proto and shared compat aliases for external callers

cargo check -p tunnel-lib: ok, cargo test -p tunnel-lib: 77 passed
