+++ tunnel-lib/src/protocol/mod.rs
@@ -1,4 +1,6 @@
+pub mod ctld;
 pub mod detect;
 pub mod driver;
 pub mod http_utils;
 pub mod rewrite;
+pub mod sniff;
