+++ tunnel-lib/src/transport/mod.rs
@@ -1,5 +1,7 @@
+pub mod accept;
 pub mod addr;
 pub mod listener;
+pub mod open_bi;
 pub mod quic;
 pub mod quinn_io;
 pub mod tcp_params;
